feat: Self-stance 固定CTAバーの背景色をカスタマイズ可能に
sticky コンテナに ss-sticky-cta-bar を追加し、装飾専用の style フィールド種別で背景色のみ編集できるようにした。
