+++ client/src/components/contents-manager/LpElementEditor.tsx
@@ -1,8 +1,10 @@
 import { DualImageElementEditor } from "@/components/contents-manager/DualImageElementEditor";
+import { FieldStyleEditor } from "@/components/contents-manager/FieldStyleEditor";
 import { LpCopyEditorSections, type WithLpFieldStyles } from "@/components/contents-manager/LpCopyEditorSections";
 import type { LpKind } from "@/lib/content-manager/cm-preview";
 import { findLpField } from "@/lib/content-manager/lp-field-registry";
 import { getFieldPath, setFieldPath } from "@/lib/content-manager/lp-field-path";
+import { patchFieldStyle } from "@/types/home-copy-style";
 
 type Props<T extends WithLpFieldStyles> = {
   kind: LpKind;
@@ -24,6 +26,22 @@ export function LpElementEditor<T extends WithLpFieldStyles>({
   const field = findLpField(slug, sectionId);
   if (!field) {
     return <p className="text-sm text-[#5C3E2A]">この要素のエディタは未設定です。</p>;
+  }
+
+  if (field.kind === "style") {
+    return (
+      <FieldStyleEditor
+        style={content.fieldStyles?.[sectionId]}
+        onStyleChange={(partial) =>
+          onChange((prev) => ({
+            ...prev,
+            fieldStyles: patchFieldStyle(prev.fieldStyles, sectionId, partial),
+          }))
+        }
+        containerOnly={field.containerOnly}
+        showHref={false}
+      />
+    );
   }
 
   if (field.kind === "image") {
