+++ client/src/components/contents-manager/TextElementEditor.tsx
@@ -1,61 +1,11 @@
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
-import { cn } from "@/lib/utils";
-import {
-  FONT_FAMILY_OPTIONS,
-  formatFontSizePx,
-  parseFontSizePx,
-  TEXT_SIZE_MAX_PX,
-  TEXT_SIZE_MIN_PX,
-  TEXT_SIZE_SLIDER_DEFAULT_PX,
-  type TextFieldStyle,
-} from "@/types/home-copy-style";
-import * as SliderPrimitive from "@radix-ui/react-slider";
+import { FieldStyleEditor } from "@/components/contents-manager/FieldStyleEditor";
+import type { TextFieldStyle } from "@/types/home-copy-style";
 
 const fieldCls = "mt-1 border-[#ffd7c3]";
 const labelCls = "text-[#3D281E]";
-
-function ColorField({
-  label,
-  value,
-  placeholder,
-  defaultPickerColor,
-  onChange,
-}: {
-  label: string;
-  value?: string;
-  placeholder: string;
-  defaultPickerColor: string;
-  onChange: (color: string) => void;
-}) {
-  return (
-    <div>
-      <Label className={labelCls}>{label}</Label>
-      <div className="flex gap-2 mt-1">
-        <Input
-          type="color"
-          value={value?.startsWith("#") ? value : defaultPickerColor}
-          onChange={(e) => onChange(e.target.value)}
-          className="h-10 w-12 p-1 border-[#ffd7c3] cursor-pointer"
-        />
-        <Input
-          value={value ?? ""}
-          onChange={(e) => onChange(e.target.value)}
-          placeholder={placeholder}
-          className={`${fieldCls} flex-1`}
-        />
-      </div>
-    </div>
-  );
-}
 
 type Props = {
   text: string;
@@ -65,49 +15,6 @@ type Props = {
   multiline?: boolean;
   rows?: number;
 };
-
-function TextSizeSlider({
-  fontSize,
-  onFontSizeChange,
-}: {
-  fontSize?: string;
-  onFontSizeChange: (px: number) => void;
-}) {
-  const savedPx = parseFontSizePx(fontSize);
-  const sliderPx = savedPx ?? TEXT_SIZE_SLIDER_DEFAULT_PX;
-  const thumbLabel = savedPx != null ? `${savedPx}px` : `${TEXT_SIZE_SLIDER_DEFAULT_PX}px`;
-
-  return (
-    <div className="mt-2 space-y-1">
-      <SliderPrimitive.Root
-        className="relative flex w-full touch-none items-center select-none py-2"
-        min={TEXT_SIZE_MIN_PX}
-        max={TEXT_SIZE_MAX_PX}
-        step={1}
-        value={[sliderPx]}
-        onValueChange={([v]) => onFontSizeChange(v)}
-        aria-label="テキストサイズ"
-      >
-        <SliderPrimitive.Track className="relative h-1.5 w-full grow overflow-hidden rounded-full bg-[#f5e6cd]">
-          <SliderPrimitive.Range className="absolute h-full rounded-full bg-[#d4844b]" />
-        </SliderPrimitive.Track>
-        <SliderPrimitive.Thumb
-          className={cn(
-            "flex h-7 min-w-[3.25rem] shrink-0 items-center justify-center rounded-full border-2 border-[#d4844b] bg-white px-2",
-            "text-xs font-semibold tabular-nums text-[#3D281E] shadow-sm",
-            "focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[#d4844b]/40",
-          )}
-        >
-          {thumbLabel}
-        </SliderPrimitive.Thumb>
-      </SliderPrimitive.Root>
-      <div className="flex justify-between text-[10px] text-[#5C3E2A]">
-        <span>{TEXT_SIZE_MIN_PX}px</span>
-        <span>{TEXT_SIZE_MAX_PX}px</span>
-      </div>
-    </div>
-  );
-}
 
 export function TextElementEditor({
   text,
@@ -137,60 +44,7 @@ export function TextElementEditor({
         )}
       </div>
 
-      <div className="grid grid-cols-2 gap-3">
-        <div className="col-span-2">
-          <Label className={labelCls}>テキストサイズ</Label>
-          <TextSizeSlider
-            fontSize={style?.fontSize}
-            onFontSizeChange={(px) => onStyleChange({ fontSize: formatFontSizePx(px) })}
-          />
-        </div>
-
-        <ColorField
-          label="テキストカラー"
-          value={style?.color}
-          placeholder="#3D281E"
-          defaultPickerColor="#3D281E"
-          onChange={(color) => onStyleChange({ color })}
-        />
-
-        <ColorField
-          label="背景色"
-          value={style?.backgroundColor}
-          placeholder="未設定（LP既定）"
-          defaultPickerColor="#3D281E"
-          onChange={(backgroundColor) => onStyleChange({ backgroundColor })}
-        />
-
-        <div>
-          <Label className={labelCls}>テキストフォント</Label>
-          <Select
-            value={style?.fontFamily ?? ""}
-            onValueChange={(v) => onStyleChange({ fontFamily: v === "__default__" ? "" : v })}
-          >
-            <SelectTrigger className={fieldCls}>
-              <SelectValue placeholder="デフォルト" />
-            </SelectTrigger>
-            <SelectContent>
-              {FONT_FAMILY_OPTIONS.map((opt) => (
-                <SelectItem key={opt.value || "__default__"} value={opt.value || "__default__"}>
-                  {opt.label}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        </div>
-
-        <div className="col-span-2">
-          <Label className={labelCls}>埋め込みURL</Label>
-          <Input
-            value={style?.href ?? ""}
-            onChange={(e) => onStyleChange({ href: e.target.value })}
-            placeholder="https://..."
-            className={fieldCls}
-          />
-        </div>
-      </div>
+      <FieldStyleEditor style={style} onStyleChange={onStyleChange} />
     </div>
   );
 }
